Update 102-fibonacci.c
Changed data types of variables x, y, and z to float. Used modifier
"%.0f" instead of "%lli" to display numbers in printf() function.

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -9,17 +9,17 @@
 int  main(void)
 {
 	int i;
-	long long x, y, z;
+	float x, y, z;
 
 	x = 1;
 	y = 2;
 
-	printf("%lli, ", x);
-	printf("%lli", y);
+	printf("%f, ", x);
+	printf("%f", y);
 	for (i = 0; i < 48; i++)
 	{
 		z = x + y;
-		printf(", %lli", z);
+		printf(", %.0f", z);
 		x = y;
 		y = z;
 	}
